Kraken Futures REST API

The listed modules and functions allow to access public and private (authenticated) endpoints of the official Kraken Futures API. All functions of these endpoints require a FuturesBaseRESTAPI instance. For accessing private endpoints, valid API keys must be generated either using the official Kraken API or the offical Kraken demo/sandbox environment. Some endpoints may not work using the demo account, other may require a higer tier rank. Please see the official Kraken documentation in the References section for futher information.

The sample outputs are for illustrative purposes only and may vary depending on the user, request and market behavior.

Kraken Futures REST API

FuturesBaseRESTAPI

Market

User

Trade

Funding

FuturesBaseRESTAPI

KrakenEx.FuturesBaseAPIModule.FuturesBaseRESTAPI — Type

FuturesBaseRESTAPI

Type that stores information about the client and can be used to access public and private endpoints of the Kraken API for Futures trading.

Fields

key – Kraken Futures API key

secret – Kraken Futures Secret key

Defualt:

BASE_URL – Kraken Futures API base url (default: "https://futures.kraken.com")

DEMO_URL – Kraken Futures API demo environment url (default: "https://demo-futures.kraken.com")

DEMO – Switch to use demo environment (default: false)

TIMEOUT – Request timeout (default: 10)

HEADERS – Default headers (default: ["User-Agent" => "KrakenEx.jl"])

Examples

FuturesBaseRESTAPI() – default, public client

FuturesBaseRESTAPI(key="the-api-key", secret="the-api-secret-key") – authenticated client for public and private requests

Market

KrakenEx.FuturesMarketModule — Module

FuturesMarketModule

Enables accessing Futures market endpoints using predefined methods.

KrakenEx.FuturesMarketModule.get_execution_events — Method

get_execution_events(
    client::FuturesBaseRESTAPI;
    before::Union{Int,String,Nothing}=nothing,
    continuation_token::Union{String,Nothing}=nothing,
    since::Union{Int,String,Nothing}=nothing,
    sort::Union{String,Nothing}=nothing,
    tradeable::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-history-market-history-get-execution-events

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_execution_events(client))
Dict{String, Any}(
    "accountUid" => "12f9bd01-d65c-4f83-9a3e-bb9464f88a35",
    "elements" => Any[], "len" => 0
)

KrakenEx.FuturesMarketModule.get_fee_schedules — Method

get_fee_schedules(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-fee-schedules-get-fee-schedules

Returns the list of maker and taker fees per asset.

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_fee_schedules(client))
Dict{String, Any}(
    "feeSchedules" => Any[
        Dict{String, Any}(
            "name" => "KF USD Multi-Collateral Fees",
            "tiers" => Any[
                Dict{String, Any}(
                    "makerFee" => 0.02,
                    "takerFee" => 0.05,
                    "usdVolume" => 0.0
                ),
                Dict{String, Any}(
                    "makerFee" => 0.015,
                    "takerFee" => 0.04,
                    "usdVolume" => 100000.0
                ),
                Dict{String, Any}(
                    "makerFee" => 0.0125,
                    "takerFee" => 0.03,
                    "usdVolume" => 1.0e6
                ), ...
            ],
            "uid" => "your-user-id"
        )
    ],
    "serverTime" => "2023-01-10T17:59:20.057Z",
    "result" => "success"
)

KrakenEx.FuturesMarketModule.get_fee_schedules_vol — Method

get_fee_schedules_vol(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-fee-schedules-get-fee-schedule-volumes

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_fee_schedules_vol(client))
Dict{String, Any}(
    "serverTime" => "2023-01-10T18:01:27.478Z",
    "volumesByFeeSchedule" => Dict{String, Any}(
        "7fc4d7c0-464f-4029-a9bb-55856d0c5247" => 0,
        "d46c2190-81e3-4370-a333-424f24387829" => 0
    ),
    "result" => "success"
)

KrakenEx.FuturesMarketModule.get_instruments — Method

get_instruments(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-instrument-details-get-instruments, https://support.kraken.com/hc/en-us/articles/360022635672-Instruments

Lists available instruments (tradeable products) and additional information.

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_instruments(client))
Dict{String,Any}(
    "result" => "success",
    "instruments" => Any[Dict{String,Any}(
        "symbol" => "pi_xbtusd",
        "type" => "futures_inverse",
        "underlying" => "rr_xbtusd",
        "tickSize" => 0.5,
        "contractSize" => 1,
        "tradeable" => true,
        "minimumTradeSize" => 1,
        "impactMidSize" => 1,
        "maxPositionSize" => 1000000,
        "openingDate" => "2022-01-01T00:00:00.000Z",
        "marginLevels" => Any[
            Dict{String,Any}("contracts" => 0, "initialMargin" => 0.02, "maintenanceMargin" => 0.01),
            Dict{String,Any}("contracts" => 500000, "initialMargin" => 0.04, "maintenanceMargin" => 0.02)
            ...
        ]
        "fundingRateCoefficient" => 8,
        "maxRelativeFundingRate" => 0.001,
        "isin" => "GB00J62YGL67",
        "contractValueTradePrecision" => 0,
        "postOnly" => false,
        "retailMarginLevels" => Any[
            Dict{String,Any}("contracts" => 0, "initialMargin" => 0.5, "maintenanceMargin" => 0.25)
        ],
        "category" => "",
        "tags" => []
    ), ...]
)

KrakenEx.FuturesMarketModule.get_instruments_status — Method

get_instruments_status(client::FuturesBaseRESTAPI; instrument::Union{String,Nothing}=nothing)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-instrument-details-get-instrument-status-list, https://docs.futures.kraken.com/#http-api-trading-v3-api-instrument-details-get-instrument-status

Returns the status of given instrument.

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_instruments_status(client, instrument="PI_XBTUSD"))
Dict{String, Any}(
    "tradeable" => "PI_XBTUSD",
    "experiencingDislocation" => false,
    "extremeVolatilityInitialMarginMultiplier" => 1,
    "priceDislocationDirection" => nothing,
    "serverTime" => "2023-01-10T18:08:20.613Z",
    "experiencingExtremeVolatility" => false,
    "result" => "success"
)

KrakenEx.FuturesMarketModule.get_leverage_preference — Method

get_leverage_preference(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-multi-collateral-get-the-leverage-setting-for-a-market

Shows the user-specific leverage preference.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_leverage_preference(client))
Dict{String,Any}(
    "result" => "success",
    "serverTime": "2022-06-28T15:01:12.762Z",
    "leveragePreferences": Any[Dict{String,Any}(
        "symbol" => "PF_XBTUSD",
        "maxLeverage" => 10.0
    )]
)

KrakenEx.FuturesMarketModule.get_ohlc — Method

get_ohlc(client::FuturesBaseRESTAPI;
    tick_type::String,
    symbol::String,
    resolution::Int64,
    from::Union{Int64,Nothing}=nothing,
    to::Union{Int64,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-charts-ohlc-get-ohlc, https://support.kraken.com/hc/en-us/articles/4403284627220-OHLC

Authenticated client not required

Arguments

symbol::String: asset of interest

tick_type::String: must be one of ["spot", "mark", "trade"]

resolution::String: must be one of ["1m", "5m", "15m", "30m", "1h", "4h", "12h", "1d", "1w"]

from::Union{Int64,Nothing}=nothing: timestamp of ohlc begin

to::Union{Int64,Nothing}=nothing: timestamp of ohlc end

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_ohlc(client,
...        tick_type="trade",
...        symbol="PI_XBTUSD",
...        resolution="1m",
...        from=1668989233,
...        to=1668999233
...    ))
Dict{String, Any}(
    "more_candles" => true,
    "candles" => Any[
        Dict{String, Any}(
            "high" => "16152.50000000000",
            "volume" => 7674,
            "time" => 1668989280000,
            "open" => "16100.00000000000",
            "low" => "16100.00000000000",
            "close" => "16144.50000000000"
        ),
        Dict{String, Any}(
            "high" => "16159.00000000000",
            "volume" => 483, "time" => 1668989340000,
            "open" => "16144.50000000000",
            "low" => "16144.50000000000",
            "close" => "16144.50000000000"
        ), ...
    ]
)

KrakenEx.FuturesMarketModule.get_order_events — Method

get_order_events(
    client::FuturesBaseRESTAPI;
    before::Union{Int,String,Nothing}=nothing,
    continuation_token::Union{String,Nothing}=nothing,
    since::Union{Int,String,Nothing}=nothing,
    sort::Union{String,Nothing}=nothing,
    tradeable::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-history-market-history-get-order-events

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_order_events(
...        client,
...        tradeable="PI_XBTUSD",
...        sort="asc",
...        before="1668989233"
...    ))
Dict{String, Any}(
    "accountUid" => "72f9bd80-d65c-4f73-9a3e-cb9464e88c35",
    "elements" => Any[
        Dict{String, Any}(
            "event" => Dict{String, Any}(
                "OrderRejected" => Dict{String, Any}(
                    "algoId" => "",
                    "order" => Dict{String, Any}(
                        "tradeable" => "PI_XBTUSD",
                        "lastUpdateTimestamp" => 1669116246449,
                        "filled" => "0",
                        "clientId" => "my_another_client_id",
                        "quantity" => "1",
                        "reduceOnly" => false,
                        "uid" => "d124183a-8c25-421b-8c75-450fb0b0efd3",
                        "direction" => "Buy",
                        "orderType" => "Limit",
                        "accountUid" => "72f9bd80-d65c-4f73-9a3e-cb9464e88c35",
                        "timestamp" => 1669116246449,
                        "limitPrice" => "1.0"
                    ),
                    "reason" => "new_user_order",
                    "orderError" => "insufficient_margin"
                )
            ),
            "uid" => "69a24b12-3a5b-470c-bac8-c65600b595d0",
            "timestamp" => 1669116246449
        )
    ],
    "len" => 1
)

KrakenEx.FuturesMarketModule.get_orderbook — Method

get_orderbook(client::FuturesBaseRESTAPI; symbol::Union{String,Nothing}=nothing)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-market-data-get-orderbook, https://support.kraken.com/hc/en-us/articles/360022839551-Order-Book

Returns the current orderbook including asks and bids of given symbol.

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_orderbook(client, symbol="PI_XBTUSD"))
Dict{String, Any}(
    "orderBook" => Dict{String, Any}(
        "asks" => Any[
            Any[17335.5, 328], Any[17339.5, 1721], Any[17340, 1770],
            Any[17341.5, 2736], Any[17343, 10379], Any[17344, 3000],
            Any[17344.5, 10366], Any[17346, 22052], Any[17346.5, 39912],
            Any[17347.5, 19052], Any[17348.5, 6000], Any[17349, 19052],
            ...
        ],
        "bids" => Any[
            ...
        ]
    ),
    "result": "success",
    "serverTime": "2020-08-27T17:03:33.196Z"
)

KrakenEx.FuturesMarketModule.get_pnl_currency_preference — Method

get_pnl_preference(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-multi-collateral-get-pnl-currency-preference-for-a-market

Returns the current asset-specific pnl preference.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_pnl_currency_preference(client))
Dict{String,Any{(
    "result" => "success",
    "serverTime" => "2022-06-28T15:04:06.710Z",
    "preferences" => Any[Dict{String,Any}(
        "symbol" => "PF_XBTUSD",
        "pnlCurrency" => "BTC"
    )]
)

KrakenEx.FuturesMarketModule.get_public_execution_events — Method

get_public_execution_events(
    client::FuturesBaseRESTAPI;
    tradeable::String,
    before::Union{Int,String,Nothing}=nothing,
    continuation_token::Union{String,Nothing}=nothing,
    since::Union{Int,String,Nothing}=nothing,
    sort::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-history-market-history-get-public-execution-events, https://support.kraken.com/hc/en-us/articles/4401755685268-Market-History-Executions

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_public_execution_events(client, tradeable="PI_XBTUSD"))
Dict{String, Any}(
    "continuationToken" => "MTY3MzM0NDAyMzMyNy83ODc0OTY5NjI4NQ==",
    "elements" => Any[Dict{String, Any}(
        "event" => Dict{String, Any}(
            "Execution" => Dict{String, Any}(
                "execution" => Dict{String, Any}(
                    "takerOrder" => Dict{String, Any}(
                        "tradeable" => "PI_XBTUSD",
                        "orderType" => "IoC",
                        "reduceOnly" => false,
                        "uid" => "b77f469e-7548-4ab5-a921-98bab50dbbb7",
                        "lastUpdateTimestamp" => 1673374561273,
                        "direction" => "Buy",
                        "timestamp" => 1673374561273,
                        "quantity" => "8999.00000000",
                        "limitPrice" => "17352.50000000"
                    ),
                    "markPrice" => "17338.91438972669",
                    "price" => "17352.5",
                    "limitFilled" => false,
                    "uid" => "dd6b4ad4-a985-4428-8689-11e7f7fe5d4c",
                    "makerOrder" => Dict{String, Any}(
                        "tradeable" => "PI_XBTUSD",
                        "orderType" => "Post",
                        "reduceOnly" => false,
                        "uid" => "82e8ff43-4dc7-4670-8f46-6a523c077ba8",
                        "lastUpdateTimestamp" => 1673374561257,
                        "direction" => "Sell",
                        "timestamp" => 1673374559677,
                        "quantity" => "27907.00000000",
                        "limitPrice" => "17352.5"
                    ),
                    "timestamp" => 1673374561273,
                    "usdValue" => "8999.00",
                    "quantity" => "8999.00000000"
                ),
                "takerReducedQuantity" => ""
            )
        ),
        "uid" => "63b6a497-b554-4af5-81bd-787f7e2bf1c7",
        "timestamp" => 1673374561273
    ), ...]
)

KrakenEx.FuturesMarketModule.get_public_mark_price_events — Method

get_public_mark_price_events(
    client::FuturesBaseRESTAPI;
    tradeable::String,
    before::Union{Int,String,Nothing}=nothing,
    continuation_token::Union{String,Nothing}=nothing,
    since::Union{Int,String,Nothing}=nothing,
    sort::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-history-market-history-get-public-mark-price-events, https://support.kraken.com/hc/en-us/articles/4401748276116-Market-History-Mark-Price

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_public_mark_price_events(client, tradeable="PI_XBTUSD"))
Dict{String, Any}(
    "continuationToken" => "MTY3MzM3NDg4NjM4MC83ODc4OTkxMDM4Nw==",
    "elements" => Any[Dict{String, Any}(
        "event" => Dict{String, Any}(
            "OrderUpdated" => Dict{String, Any}(
                "reducedQuantity" => "0",
                "oldOrder" => Dict{String, Any}(
                    "tradeable" => "PI_XBTUSD",
                    "orderType" => "Post",
                    "reduceOnly" => false,
                    "uid" => "28e2dc51-e473-4621-ad30-88d289e2258e",
                    "lastUpdateTimestamp" => 1673374894474,
                    "direction" => "Sell",
                    "timestamp" => 1673354694830,
                    "quantity" => "6000",
                    "limitPrice" => "17384.0"
                ),
                "newOrder" => Dict{String, Any}(
                    "tradeable" => "PI_XBTUSD",
                    "orderType" => "Post",
                    "reduceOnly" => false,
                    "uid" => "28e2dc51-e473-4621-ad30-88d289e2258e",
                    "lastUpdateTimestamp" => 1673374894696,
                    "direction" => "Sell",
                    "timestamp" => 1673354694830,
                    "quantity" => "6000",
                    "limitPrice" => "17376.0"
                ), "reason" => "edited_by user"
            )
        ),
        "uid" => "fc6fd47a-303b-45fc-a187-d6345b664560",
        "timestamp" => 1673374894696
    ), ...]
)

KrakenEx.FuturesMarketModule.get_public_order_events — Method

get_public_order_events(
    client::FuturesBaseRESTAPI;
    tradeable::String,
    before::Union{Int,String,Nothing}=nothing,
    continuation_token::Union{String,Nothing}=nothing,
    since::Union{Int,String,Nothing}=nothing,
    sort::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-history-market-history-get-public-order-events, https://support.kraken.com/hc/en-us/articles/4401755906452-Market-History-Orders

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_public_order_events(client, tradeable="PI_XBTUSD"))
Dict{String, Any}(
    "continuationToken" => "MTY3MzM3NDczMjg2MC83ODc4OTc0NDU2OA==",
    "elements" => Any[
        Dict{String, Any}(
            event" => Dict{String, Any}(
                "OrderPlaced" => Dict{String, Any}(
                    "reducedQuantity" => "",
                    "order" => Dict{String, Any}(
                        "tradeable" => "PI_XBTUSD",
                        "orderType" => "Post",
                        "reduceOnly" => false,
                        "uid" => "68fb6c2e-5f85-44b9-a6f2-aae9ac6cd0c5",
                        "lastUpdateTimestamp" => 1673374745391,
                        "direction" => "Buy",
                        "timestamp" => 1673374745391,
                        "quantity" => "118",
                        "limitPrice" => "17167.0"
                    ),
                    "reason" => "new_user_order"
                )
            ),
            "uid" => "58733df7-f573-4ddd-96c5-bf4978b6568d",
            "timestamp" => 1673374745391
        ), Dict{String, Any}(
            "event" => Dict{String, Any}(
                "OrderUpdated" => Dict{String, Any}(
                    "reducedQuantity" => "0",
                    "oldOrder" => Dict{String, Any}(
                        "tradeable" => "PI_XBTUSD",
                        "orderType" => "Post",
                        "reduceOnly" => false,
                        "uid" => "54de25df-f2e4-49d1-8281-c6a867fa2e8e",
                        "lastUpdateTimestamp" => 1673374745228,
                        "direction" => "Sell",
                        "timestamp" => 1673352860129,
                        "quantity" => "300000",
                        "limitPrice" => "17381.0"
                    ),
                    "newOrder" => Dict{String, Any}(
                        "tradeable" => "PI_XBTUSD",
                        "orderType" => "Post",
                        "reduceOnly" => false,
                        "uid" => "54de25df-f2e4-49d1-8281-c6a867fa2e8e",
                        "lastUpdateTimestamp" => 1673374745384,
                        "direction" => "Sell",
                        "timestamp" => 1673352860129,
                        "quantity" => "300000",
                        "limitPrice" => "17383.0"
                    ),
                    "reason" => "edited_by user"
                )
            ),
            "uid" => "f741b4a5-44d6-45a7-b287-77eba4a167b7",
            "timestamp" => 1673374745384
        ), ...
    ]
)

KrakenEx.FuturesMarketModule.get_resolutions — Method

get_resolutions(client::FuturesBaseRESTAPI; tick_type::String, tradeable::String)

Kraken Docs: https://docs.futures.kraken.com/#http-api-charts-ohlc-get-resolutions

Returns the list of available resolutions.

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia>  println(get_resolutions(client, tradeable="PI_XBTUSD", tick_type="spot"))
Any["1d", "5m", "15m", "30m", "4h", "1m", "1h", "1w", "12h"]

KrakenEx.FuturesMarketModule.get_tick_types — Method

get_tick_types(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-charts-ohlc-get-tick-types

Returns the available tick types.

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_tick_types(client))
Any["mark", "spot", "trade"]

KrakenEx.FuturesMarketModule.get_tickers — Method

get_tickers(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-market-data-get-tickers, https://support.kraken.com/hc/en-us/articles/360022839531-Tickers

Returns the available ticker for all available Futures symbols.

Authenticated client not required

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_tickers(client))
Dict{String, Any}(
    "serverTime" => "2023-01-10T18:04:18.210Z",
    "tickers" => Any[
        Dict{String, Any}(
            "fundingRate" => 3.6841376588881e-5,
            "lastSize" => 2.6, "bid" => 35.48,
            "fundingRatePrediction" => -0.000541681480000012,
            "lastTime" => "2023-01-10T11:42:04.518Z",
            "pair" => "COMP:USD",
            "bidSize" => 14,
            "volumeQuote" => 331.9023,
            "last" => 35.419,
            "suspended" => false,
            "askSize" => 17,
            "postOnly" => false,
            "symbol" => "pf_compusd",
            "openInterest" => 617.5,
            "markPrice" => 35.527,
            "ask" => 35.575,
            "tag" => "perpetual",
            "vol24h" => 9.3,
            "open24h" => 36,
            "indexPrice" => 35.524
        ), Dict{String, Any}(
            "lastSize" => 250,
            "bid" => 80.27,
            "lastTime" => "2023-01-10T18:03:31.516Z",
            "pair" => "LTC:USD",
            ...
        ), ...
    ]
)

KrakenEx.FuturesMarketModule.get_trade_history — Method

get_trade_history(client::FuturesBaseRESTAPI;
    symbol::Union{String,Nothing}=nothing,
    lastTime::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-market-data-get-trade-history, https://support.kraken.com/hc/en-us/articles/360022839511-History

Lists historical trades of this user.

Authenticated client not required

Note

lastTime must be in the format like: 2020-08-27T17:03:33.196Z

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_trade_history(client, symbol="PI_XBTUSD"))
Dict{String, Any}(
    "history" => Any[
        Dict{String, Any}(
            "price" => 17308.5,
            "time" => "2023-01-10T16:58:58.667Z",
            "side" => "sell",
            "uid" => "15e8ea9f-6ccf-4545-b29f-18b9e8f93949",
            "size" => 1730,
            "type" => "fill",
            "trade_id" => 100
        ), Dict{String, Any}(
            "price" => 17289.5,
            "time" => "2023-01-10T17:01:13.391Z",
            "side" => "buy",
            "uid" => "1475a325-0d14-46bd-9fb7-0deab5746e1c",
            "size" => 4524,
            "type" => "fill",
            "trade_id" => 99
        ), ...
    ],
    "result": "success",
    "serverTime": "2022-08-27T17:03:33.196Z"
)

KrakenEx.FuturesMarketModule.get_tradeable_products — Method

get_tradeable_products(client::FuturesBaseRESTAPI; tick_type::String)

Kraken Docs: https://docs.futures.kraken.com/#http-api-charts-ohlc-get-tradeable-products

Returns a list of available products for trading.

Authenticated client not required

Attributes

tick_type::String: must be one of ["spot", "mark", "trade"]

Examples

julia> client = FuturesBaseRESTAPI()
julia> println(get_tradeable_products(client, tick_type="mark"))
Any[
    "FI_ETHUSD_210924", "PF_LUNA2USD", "FI_LTCUSD_221230", "PF_UNIUSD",
    "FI_XBTUSD_230331", "FI_BCHUSD_211126", "PI_ETHUSD", "FI_ETHUSD_220527",
    ...
]

KrakenEx.FuturesMarketModule.get_trigger_events — Method

get_trigger_events(
    client::FuturesBaseRESTAPI;
    before::Union{Int,String,Nothing}=nothing,
    continuation_token::Union{String,Nothing}=nothing,
    since::Union{Int,String,Nothing}=nothing,
    sort::Union{String,Nothing}=nothing,
    tradeable::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-history-market-history-get-trigger-events

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_trigger_events(
...        client,
...        tradeable="PI_XBTUSD",
...        sort="desc",
...        before="1668989233"
...    ))
Dict{String, Any}(
    "accountUid" => "12f2b283-af1a-a173-aaae-ce9a64e122abc1",
    "elements" => Any[],
    "len" => 0
)

KrakenEx.FuturesMarketModule.set_leverage_preference — Method

set_leverage_preference(client::FuturesBaseRESTAPI; symbol::String, maxLeverage::Union{Float64,Nothing}=nothing)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-multi-collateral-set-the-leverage-setting-for-a-market

Enables setting the leverage preference.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(set_leverage_preference(client, symbol="PI_XBTUSD", maxLeverage=1))
Dict{String,Any}(
    "result" => "success",
    "serverTime" => "2022-06-28T14:48:58.711Z"
)

KrakenEx.FuturesMarketModule.set_pnl_currency_preference — Method

set_pnl_currency_preference(client::FuturesBaseRESTAPI; symbol::String, pnlPreference::String)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-multi-collateral-set-pnl-currency-preference-for-a-market

Enables editing the current asset-specific pnl preference.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(set_pnl_currency_preference(client, symbol="PI_XBTUSD", pnlPreference="XBT"))
Dict{String,Any}(
    "result" => "success",
    "serverTime" => "2022-06-28T14:48:58.711Z"
)

User

KrakenEx.FuturesUserModule — Module

FuturesUserModule

Enables accessing Futures user endpoints using predefined methods.

KrakenEx.FuturesUserModule.get_account_log — Method

get_account_log(client::FuturesBaseRESTAPI;
    before::Union{Int,String,Nothing}=nothing,
    since::Union{Int,String,Nothing}=nothing,
    count::Union{Int,String,Nothing}=nothing,
    from::Union{Int,String,Nothing}=nothing,
    to::Union{Int,String,Nothing}=nothing,
    sort::Union{String,Nothing}=nothing,
    info::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-history-account-log

Enables querying the log of specific events.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_account_log(
...        client, 
...        before="1604937694000", 
...        info="futures liquidation"
...    ))
Dict{String, Any}(
    "accountUid" => "XXX-XXX-XXX--XXXX", 
    "logs" => Any[]
)

KrakenEx.FuturesUserModule.get_account_log_csv — Method

get_account_log_csv(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-history-account-log-get-recent-account-log-csv

Enables requesting the account log to save it as for example as csv.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> log = get_account_log_csv(client)
julia> open("myAccountLog.csv", "w") do io
...        write(io, log)
...    end

KrakenEx.FuturesUserModule.get_notificatios — Method

get_notificatios(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-general-get-notifications

Returns information defined by Kraken, e.g. "PostOnly mode at 01.01.2023 00:00:00"

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_notificatios(client))
Dict{String, Any}(
    "serverTime" => "2023-01-10T18:33:28.582Z", 
    "notifications" => Any[], 
    "result" => "success"
)

KrakenEx.FuturesUserModule.get_open_orders — Method

get_open_orders(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-account-information-get-open-orders

Lists the open orders of this user.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_open_orders(client))
Dict{String, Any}(
    "openOrders" => Any[], 
    "serverTime" => "2023-01-10T18:31:07.027Z", 
    "result" => "success"
)

KrakenEx.FuturesUserModule.get_open_positions — Method

get_open_positions(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-account-information-get-open-positions

Lists the open positions of this user.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_open_positions(client))
Dict{String, Any}(
    "openPositions" => Any[],
    "serverTime" => "2023-01-10T18:31:45.903Z", 
    "result" => "success
)

KrakenEx.FuturesUserModule.get_subaccounts — Method

get_subaccounts(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-account-information-get-subaccounts

Lists the available subaccounts.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_subaccounts(client))
Dict{String, Any}(
    "subaccounts" => Any[], 
    "serverTime" => "2023-01-10T18:32:03.993Z", 
    "masterAccountUid" => "XXX-XXX-XXX-XXX", 
    "result" => "success"
)

KrakenEx.FuturesUserModule.get_unwindqueue — Method

get_unwindqueue(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-account-information-get-position-percentile-of-unwind-queue

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_unwindqueue(client))
Dict{String, Any}(
    "queue" => Any[], 
    "serverTime" => "2023-01-10T18:33:00.647Z", 
    "result" => "success"
)

KrakenEx.FuturesUserModule.get_wallets — Method

get_wallets(client::FuturesBaseRESTAPI)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-account-information-get-wallets

Lists the user-specifc wallets.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_wallets(client))
ict{String, Any}(
    "serverTime" => "2023-01-10T18:29:00.251Z", 
    "accounts" => Dict{String, Any}(
        "fv_xrpxbt" => Dict{String, Any}(
            "currency" => "xbt", 
            "marginRequirements" => Dict{String, Any}(
                "tt" => 0.0, "mm" => 0.0, "lt" => 0.0, "im" => 0.0
            ), 
            "auxiliary" => Dict{String, Any}(
                "pv" => 0.0, "pnl" => 0.0, "funding" => 0.0, "af" => 0.0, "usd" => 0
            ), 
            "balances" => Dict{String, Any}("xbt" => 0.0), 
            "triggerEstimates" => Dict{String, Any}(
                "tt" => 0, "mm" => 0, "lt" => 0, "im" => 0
            ), 
            "type" => "marginAccount"
        ), 
        "fi_xbtusd" => Dict{String, Any}(
            "currency" => "xbt", 
            "marginRequirements" => Dict{String, Any}(
                "tt" => 0.0, "mm" => 0.0, "lt" => 0.0, "im" => 0.0
            ), 
            "auxiliary" => Dict{String, Any}(
                "pv" => 0.0, "pnl" => 0.0, "funding" => 0.0, "af" => 0.0, "usd" => 0
            ), 
            "balances" => Dict{String, Any}("xbt" => 0.0),
            "triggerEstimates" => Dict{String, Any}(
                "tt" => 0, "mm" => 0, "lt" => 0, "im" => 0
            ), "type" => "marginAccount"
        ), ...
    )
)

Trade

KrakenEx.FuturesTradeModule — Module

FuturesTradeModule

Enables accessing Futures trade endpoints using predefined methods.

KrakenEx.FuturesTradeModule.cancel_all_orders — Method

cancel_all_orders(client::FuturesBaseRESTAPI; symbol::Union{String,Nothing}=nothing)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-order-management-cancel-all-orders

Cancels all active orders of a user.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(cancel_all_orders(client))
Dict{String,Any}(
    "result" => "success",
    "cancelStatus" => Dict{String,Any}(
        "receivedTime" => "2019-08-01T15:57:37.518Z",
        "cancelOnly" => "all",
        "status" => "cancelled",
        "cancelledOrders": Any[
            Dict{String,Any}("order_id" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX"),
            Dict{String,Any}("order_id" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX"),
            ...
        ],
        "orderEvents" => Any[Dict{String,Any}(
            "uid" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
            "order" => Dict{String,Any}(
                "orderId" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
                "cliOrdId" => null,
                "type" => "stp",
                "symbol" => "pi_xbtusd",
                "side" => "buy",
                "quantity" => 2001,
                "filled" => 0,
                "limitPrice" => 11031,
                "stopPrice" => 11031,
                "reduceOnly" => false,
                "timestamp" => "2022-08-01T15:57:19.482Z"
            ),
            "type": "CANCEL"
        ), ...]
    )
)

KrakenEx.FuturesTradeModule.cancel_order — Method

cancel_order(client::FuturesBaseRESTAPI;
    order_id::Union{String,Nothing}=nothing,
    cliOrdId::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-order-management-cancel-order

Cancels an order by order_id or cliOrdId.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(cancel_order(client, order_id="XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX"))
Dict{String,Any}(
    "result" => "success",
    "cancelStatus" => Dict{String,Any}(
        "status" => "cancelled",
        "order_id" => ""XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX,
        "cancelOnly" => "all",
        "receivedTime" => "2019-08-01T15:57:37.518Z",
        "orderEvents" => Any[Dict{String,Any}(
            "uid" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
            "order" => Dict{String,Any}(
                "orderId" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
                "cliOrdId" => null,
                "type" => "stp",
                "symbol" => "pi_xbtusd",
                "side" => "buy",
                "quantity" => 2001,
                "filled" => 0,
                "limitPrice" => 11031,
                "stopPrice" => 11031,
                "reduceOnly" => false,
                "timestamp" => "2022-08-01T15:57:19.482Z"
            ),
            "type": "CANCEL"
        ), ...]
    )
)

KrakenEx.FuturesTradeModule.create_batch_order — Method

create_batch_order(client::FuturesBaseRESTAPI; batchorder_list::Vector{Dict{String,Any}})

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-order-management-batch-order-management

Creates a batch of order instructions to create, edit or even cancel multiple orders at the same time. Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(create_batch_order(
...        client,
...        batchorder_list=[
...            Dict{String,Any}(
...                "order" => "send",
...                "order_tag" => "1",
...                "orderType" => "lmt",
...                "symbol" => "PI_XBTUSD",
...                "side" => "buy",
...                "size" => 2,
...                "limitPrice" => 15200.0,
...                "cliOrdId" => "my-another-client-id"
...            ),
...            Dict{String,Any}(
...                "order" => "send",
...                "order_tag" => "2",
...                "orderType" => "stp",
...                "symbol" => "PI_XBTUSD",
...                "side" => "buy",
...                "size" => 1,
...                "limitPrice" => 15100.0,
...                "stopPrice" => 15090.0,
...            ),
...            Dict{String,Any}(
...                "order" => "send",
...                "order_tag" => "3",
...                "orderType" => "stp",
...                "symbol" => "PI_XBTUSD",
...                "side" => "buy",
...                "size" => 3,
...                "limitPrice" => 15100.0,
...                "stopPrice" => 14900.0,
...            ),
...            Dict{String,Any}(
...                "order" => "cancel",
...                "order_id" => "f35a61dd-8a30-4d5f-a574-b5593ef0c050",
...            ),
...            Dict{String,Any}(
...                "order" => "cancel",
...                "cliOrdId" => "my_client_id1234"
...            )
...        ]))
...
julia> # please see the linked Kraken Docs for response examples

KrakenEx.FuturesTradeModule.create_order — Method

create_order(client::FuturesBaseRESTAPI;
    orderType::String,
    side::String,
    size::Union{Float64,Int,String},
    symbol::String,
    cliOrdId::Union{String,Nothing}=nothing,
    limitPrice::Union{String,Float64,Int,Nothing}=nothing,
    stopPrice::Union{String,Float64,Int,Nothing}=nothing,
    reduceOnly::Union{Bool,Nothing}=nothing,
    triggerSignal::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-order-management-send-order

Creates/places an order.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(create_order(client,
...         orderType="lmt",
...         side="buy",
...         size=1,
...         limitPrice=16000.0,
...         symbol="PI_XBTUSD",
...     ))
Dict{String,Any}(
    "result" => "success",
    "sendStatus" => Dict{String,Any}(
        "order_id" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
        "status" => "placed",
        "receivedTime" => "2022-12-11T17:17:33.888Z",
        "orderEvents": Any[Dict{String,Any}(
            "executionId" => "XXXXX0XX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
            "price" => 17000.0,
            "amount" => 1,
            "orderPriorEdit" => null,
            "orderPriorExecution" => Dict{String,Any}(
                    "orderId" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
                    "cliOrdId" => null,
                    "type" => "lmt",
                    "symbol" => "pi_xbtusd",
                    "side" => "buy",
                    "quantity" => 1,
                    "filled" => 0,
                    "limitPrice" => 16000.0,
                    "reduceOnly" => false,
                    "timestamp" => "2012-12-11T17:17:33.888Z",
                    "lastUpdateTimestamp" => "2012-12-11T17:17:33.888Z"
            ),
            "takerReducedQuantity" => null,
            "type" => "EXECUTION"
        )]
    ),
    "serverTime": "2019-12-11T17:17:33.888Z"
)

KrakenEx.FuturesTradeModule.dead_mans_switch — Method

dead_mans_switch(client::FuturesBaseRESTAPI; timeout::Int=60)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-order-management-dead-man-39-s-switch

Enables the dead mans switch to prevent unwanted behaviour e.g. during system instabilities. Set timeout=0 ro reset.

Authenticated client required

Set timeout to 0 to reset the death man switch.

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(dead_mans_switch(client, timeout=60))
Dict{String,Any}(
    "result" => "success",
    "status" => Dict{String,Any}(
        "currentTime" => "2018-06-19T16:51:23.839Z",
        "triggerTime" => "60"
    ),
    "serverTime" => "2018-06-19T16:51:23.839Z"
)
julia> println(dead_mans_switch(client, timeout=0))
Dict{String,Any}(
    "result" => "success",
    "status" => Dict{String,Any}(
        "currentTime" => "2018-06-19T16:51:23.839Z",
        "triggerTime" => "0"
    ),
    "serverTime" => "2018-06-19T16:51:23.839Z"
)

KrakenEx.FuturesTradeModule.edit_order — Method

edit_order(client::FuturesBaseRESTAPI;
    orderId::Union{String,Nothing}=nothing,
    cliOrdId::Union{String,Nothing}=nothing,
    limitPrice::Union{String,Float64,Int,Nothing}=nothing,
    size::Union{String,Float64,Int,Nothing}=nothing,
    stopPrice::Union{String,Float64,Int,Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-order-management-edit-order

Enables editing placed orders which are still open.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(edit_order(client,
...        orderId="XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
...        size=2,
...        limitPrice=16000,
...    ))
Dict{String,Any}(
    "result" => "success",
    "serverTime" => "2022-09-05T16:47:47.521Z",
    "editStatus" => Dict{String,Any}(
        "status" => "edited",
        "orderId" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
        "receivedTime" => "2022-09-05T16:47:47.521Z",
        "orderEvents" => Any[
            Dict{String,Any}(
                "old" => Dict{String,Any}(
                    "orderId" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
                    "cliOrdId" => null,
                    "type" => "lmt",
                    "symbol" => "pi_xbtusd",
                    "side" => "buy",
                    "quantity" => 1,
                    "filled": => 0,
                    "limitPrice" => 17000.0,
                    "reduceOnly" => false,
                    "timestamp" => "2022-09-05T16:41:35.173Z",
                    "lastUpdateTimestamp" => "2019-09-05T16:41:35.173Z"
                )
                "new" => Dict{String,Any}(
                    "orderId" => "XXXXXXXX-XXX-XXXX-XXXX-XXXXXXXXXXXX",
                    "cliOrdId" => null,
                    "type" => "lmt",
                    "symbol" => "pi_xbtusd",
                    "side" => "buy",
                    "quantity" => 2,
                    "filled" => 0,
                    "limitPrice" => 16000.0,
                    "reduceOnly" => false,
                    "timestamp" => "2022-09-05T16:41:35.173Z",
                    "lastUpdateTimestamp" => "2019-09-05T16:47:47.519Z"
                ),
                "reducedQuantity": null,
                "type": "EDIT"
            )
        ]
    )
)

KrakenEx.FuturesTradeModule.get_fills — Method

get_fills(client::FuturesBaseRESTAPI; lastFillTime::Union{String,Nothing}=nothing)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-historical-data-get-your-fills

Returns information about the user-specific fills of all Futures contracts.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_fills(client))
Dict{String,Any}(
    "result" => "success",
    "fills" => Any[Dict{String,Any}(
        "fill_id" => "3d57ed09-fbd6-44f1-8e8b-b10e551c5e73",
        "symbol" => "pi_xbtusd",
        "side" => "buy",
        "order_id" => "693af756-055e-47ef-99d5-bcf4c456ebc5",
        "size" => 5490,
        "price" => 9400,
        "fillTime" => "2020-07-22T13:37:27.077Z",
        "fillType" => "maker"
    ), ...],
    "serverTime" => "2020-07-22T13:44:24.311Z"
)

KrakenEx.FuturesTradeModule.get_orders_status — Method

get_orders_status(client::FuturesBaseRESTAPI;
    orderIds::Union{Vector{String},Nothing}=nothing,
    cliOrdIds::Union{Vector{String},Nothing}=nothing
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-order-management-get-the-current-status-for-specific-orders

Get information about (a) specific order(s) using orderIds or cliOrdIds.

Authenticated client required

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_orders_status(client, orderIds=["orderid1", "orderid2"]))
Dict{String,Any}(
    "orders" => Any[Dict{String,Any}
      "error" => "MARKET_SUSPENDED",
      "order" => {
        "cliOrdId" => "someniceid",
        "filled" => 0.1,
        "lastUpdateTimestamp": "2021-08-26T17:03:33.196Z",
        "limitPrice": 2,
        "orderId": "orderid1",
        "priceTriggerOptions": {
          "triggerPrice": 1,
          "triggerSide": "TRIGGER_ABOVE",
          "triggerSignal": "MARK_PRICE"
        },
        "quantity": 5,
        "reduceOnly": false,
        "side": "buy",
        "symbol": "DOTUSD",
        "timestamp": "2021-08-27T17:03:33.196Z",
        "type": "TRIGGER_ORDER"
      },
      "status": "ENTERED_BOOK",
      "updateReason": "LOADING_MARKET"
    }
  ],
  "result": "success",
  "serverTime": "2022-08-27T17:03:33.196Z"
)

Funding

KrakenEx.FuturesFundingModule — Module

FuturesFundingModule

Enables accessing Futures funding endpoints using predefined methods.

KrakenEx.FuturesFundingModule.get_historical_funding_rates — Method

get_historical_funding_rates(client::FuturesBaseRESTAPI; symbol::String)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-historical-funding-rates-historicalfundingrates

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> println(get_historical_funding_rates(client, symbol="PI_XBTUSD"))
Dict{String, Any}(
    "rates" => Any[
        Dict{String, Any}(
            "fundingRate" => 1.0327058177e-8,
            "timestamp" => "2018-08-31T16:00:00.000Z",
            "relativeFundingRate" => 7.182407e-5
        ),
        Dict{String, Any}(
            "fundingRate" => -1.2047162502e-8,
            "timestamp" => "2018-08-31T20:00:00.000Z",
            "relativeFundingRate" => -8.4873103125e-5
        ),
        Dict{String, Any}(
            "fundingRate" => -9.645113378e-9,
            "timestamp" => "2018-09-01T00:00:00.000Z",
            "relativeFundingRate" => -6.76651e-5
        ),
        Dict{String, Any}(
            "fundingRate" => -8.028122964e-9,
            "timestamp" => "2018-09-01T04:00:00.000Z",
            "relativeFundingRate" => -5.66897875e-5
        ), ...
    ]
)

KrakenEx.FuturesFundingModule.initiate_subccount_transfer — Method

initiate_subccount_transfer(
    client::FuturesBaseRESTAPI;
    amount::Union{String,Float64,Int},
    fromAccount::String,
    toAccount::String,
    unit::String,
    fromUser::String,
    toUser::String
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-transfers-initiate-sub-account-transfer

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> initiate_subccount_transfer(
...        client,
...        amount=200,
...        fromAccount="SomeAccount",
...        toAccount="ToAnotherAccount",
...        fromUser="Marianne",
...        toUser="Perl",
...        unit="XBT"
...    )

KrakenEx.FuturesFundingModule.initiate_wallet_transfer — Method

initiate_wallet_transfer(
    client::FuturesBaseRESTAPI;
    amount::Union{String,Float64,Int},
    fromAccount::String,
    toAccount::String,
    unit::String
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-transfers-initiate-wallet-transfer

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> initiate_wallet_transfer(client,
...        amount=100,
...        fromAccount="some-cash-or-margin-account",
...        toAccount="another-cash-or-margin-account",
...        unit="the-currency-unit-to-transfer"
...    )

KrakenEx.FuturesFundingModule.initiate_withdrawal_to_spot_wallet — Method

initiate_withdrawal_to_spot_wallet(
    client::FuturesBaseRESTAPI;
    amount::Union{String,Int,Float64},
    currency::String,
    sourceWallet::Union{String,Nothing}
)

Kraken Docs: https://docs.futures.kraken.com/#http-api-trading-v3-api-transfers-initiate-withdrawal-to-spot-wallet

Examples

julia> client = FuturesBaseRESTAPI(key="api-key", secret="api-secret")
julia> initiate_withdrawal_to_spot_wallet(
...       client,
...       amount=200,
...       currency="XBT",
...       sourceWallet="Cross-Margin"
...    )