Kraken Spot REST API

The listed modules and functions allow to access public and private (authenticated) endpoints of the official Kraken Spot API. All functions of these endpoints require a SpotBaseRESTAPI instance.

The sample outputs are for illustrative purposes only and may vary depending on the user, request, and market behavior.

Kraken Spot REST API

SpotBaseRESTAPI

Market

User

Trade

Funding

Staking

SpotBaseRESTAPI

KrakenEx.SpotBaseAPIModule.SpotBaseRESTAPI — Type

SpotBaseRESTAPI

Type that stores information about the client and can be used to access public and private endpoints of the Kraken API for Spot trading.

Fields

key – Kraken Spot API key

secret – Kraken Spot Secret key

Defualt:

BASE_URL – Kraken Spot API base url (default: "https://api.kraken.com")

API_V – Kraken Spot API version (default: 0)

TIMEOUT – Request timeout (default: 10)

HEADERS – Default headers {default: ["User-Agent" => "KrakenEx.jl"]}

Examples

SpotBaseRESTAPI() – default, public client

SpotBaseRESTAPI(key="the-api-key", secret="the-api-secret-key") – authenticated client for public and private requests

Market

KrakenEx.SpotMarketModule.get_assets — Method

get_assets(client::SpotBaseRESTAPI; assets::Union{Vector{String},String,Nothing}=nothing, aclass::Union{String,Nothing}=nothing)

Kraken Docs: https://docs.kraken.com/rest/#operation/getAssetInfo

Returns a list of available Spot assets as well as additional descriptions.

Authenticated client not required

Examples

julia> client = SpotBaseRESTAPI()
julia> println(get_assets(client))
Dict{String, Any}(
    "MOVR" => Dict{String, Any}(
        "display_decimals" => 5, 
        "decimals" => 10, 
        "aclass" => "currency", 
        "status" => "enabled", 
        "altname" => "MOVR"
    ), "SXP" => Dict{String, Any}(
        "display_decimals" => 5, 
        "decimals" => 10, 
        "aclass" => "currency", 
        "status" => "enabled", 
        "altname" => "SXP"
    ), "PHA" => Dict{String, Any}(
        "display_decimals" => 5, 
        "decimals" => 10, 
        "aclass" => "currency", 
        "status" => "enabled", 
        "altname" => "PHA"
    ), ...
)

KrakenEx.SpotMarketModule.get_ohlc — Method

get_ohlc(client::SpotBaseRESTAPI; pair::String, interval::Union{String,Nothing}=nothing, since::Union{Int64,Nothing}=nothing)

Kraken Docs: https://docs.kraken.com/rest/#operation/getOHLCData

Returns information about the open, high, low, close, vwap, and volume of a given pair.

Authenticated client not required

Examples

julia> client = SpotBaseRESTAPI()
julia> get_ohlc(client, pair="XBTUSD")
Dict{String, Any}(
    "XXBTZUSD" => Any[
        Any[1673301660, "17176.5", "17176.5", "17176.4", "17176.5", "17176.4", "0.00499447", 3], 
        Any[1673301720, "17176.5", "17176.5", "17167.2", "17168.1", "17168.3", "3.57307618", 45], 
        Any[1673301780, "17168.0", "17168.1", "17167.9", "17168.1", "17167.9", "0.10257254", 16], 
        Any[1673301840, "17168.0", "17176.5", "17168.0", "17176.5", "17176.2", "1.06103454", 19],
        Any[1673301900, "17176.4", "17176.5", "17176.4", "17176.5", "17176.4", "0.20158908", 4], 
        Any[1673301960, "17176.5", "17176.5", "17176.4", "17176.5", "17176.4", "0.04299818", 7], 
        ...
    ]
)

KrakenEx.SpotMarketModule.get_order_book — Method

get_order_book(client::SpotBaseRESTAPI; pair::String, count::Union{Int64,Nothing}=nothing)

Kraken Docs: https://docs.kraken.com/rest/#operation/getOrderBook

Returns the current asks and bids (orderbook) of a given currency pair.

Authenticated client not required

Examples

julia> client = SpotBaseRESTAPI()
julia> get_order_book(client, pair="XBTUSD")
Dict{String, Any}(
    "XXBTZUSD" => Dict{String, Any}(
        "asks" => Any[
            Any["17249.60000", "0.750", 1673344930], Any["17249.80000", "0.038", 1673344927], 
            Any["17249.90000", "0.067", 1673344930], Any["17250.10000", "0.001", 1673344930], 
            Any["17250.50000", "4.348", 1673344903], Any["17251.10000", "0.053", 1673344928], 
            Any["17251.30000", "1.157", 1673344928], Any["17251.50000", "0.807", 1673344930], 
            Any["17252.40000", "0.369", 1673344921], Any["17252.60000", "0.250", 1673344857], 
            ...
        ],
        "bids" => Any[
            Any["17249.50000", "0.001", 1673344924], 
            Any["17249.40000", "0.001", 1673344916], 
            Any["17249.30000", "0.001", 1673344930], 
            Any["17249.20000", "0.001", 1673344927], 
            Any["17249.10000", "0.001", 1673344930], 
            ...
        ]
    )
)

KrakenEx.SpotMarketModule.get_recend_spreads — Method

get_recend_spreads(client::SpotBaseRESTAPI; pair::String, since::Union{Int64,Nothing}=nothing)

Kraken Docs: https://docs.kraken.com/rest/#operation/getRecentSpreads

Returns the recend spreads of a currency pair (symbol).

Authenticated client not required

Examples

julia> client = SpotBaseRESTAPI()
julia> println(get_recend_spreads(client, pair="XBTUSD"))
Dict{String, Any}(
    "XXBTZUSD" => Any[
        Any[1673363656, "17346.40000", "17348.20000"], 
        Any[1673363657, "17346.40000", "17347.90000"],
        Any[1673363657, "17346.80000", "17347.90000"], 
        Any[1673363660, "17346.80000", "17347.60000"], 
        Any[1673363660, "17346.40000", "17347.60000"], 
        Any[1673363660, "17346.50000", "17347.60000"], 
        ...
    ], 
    "last" => 1673363958)

KrakenEx.SpotMarketModule.get_recent_trades — Method

get_recent_trades(client::SpotBaseRESTAPI; pair::String, since::Union{Int64,Nothing}=nothing)

Kraken Docs: https://docs.kraken.com/rest/#operation/getRecentTrades

Returns the recent trades of a given symbol (currency pair).

Authenticated client not required

Examples

julia> client = SpotBaseRESTAPI()
julia> println(get_recent_trades(client, pair="XBTUSD"))
Dict{String, Any}(
    "XXBTZUSD" => Any[
        Any["17265.50000", "0.00515323", 1.673362549987055e9, "b", "l", "", 54454372], 
        Any["17265.50000", "0.01203124", 1.6733625510070226e9, "b", "l", "", 54454373], 
        Any["17266.80000", "0.08087410", 1.673362551007105e9, "b", "l", "", 54454374], 
        Any["17266.90000", "0.56059466", 1.6733625510071838e9, "b", "l", "", 54454375], 
        Any["17267.00000", "0.00900000", 1.6733625512079215e9, "b", "l", "", 54454376], 
        Any["17267.00000", "0.00610000", 1.6733625522878373e9, "b", "l", "", 54454377], 
        Any["17267.00000", "0.00208447", 1.6733625534015026e9, "b", "l", "", 54454378], 
        Any["17268.50000", "0.01211553", 1.6733625534015663e9, "b", "l", "", 54454379],
        Any["17268.50000", "0.00506894", 1.6733625536050434e9, "b", "l", "", 54454380], 
        ...
    ], 
    "last" => "1673363844906015910"
)

KrakenEx.SpotMarketModule.get_server_time — Method

get_server_time(client::SpotBaseRESTAPI)

Returns the server time of the Kraken Cryptocurrency Exchange API

Examples

julia> client = SpotBaseRESTAPI()
julia> println(get_server_time(client))
1673344450

KrakenEx.SpotMarketModule.get_system_status — Method

get_system_status(client::SpotBaseRESTAPI)

Returns the system status of the Kraken API

Authenticated client not required

Examples

julia> client = SpotBaseRESTAPI()
julia> println(get_system_status(client))
Dict{String, Any}("status" => "online", "timestamp" => "2023-01-10T15:20:20Z")

KrakenEx.SpotMarketModule.get_ticker — Method

get_ticker(client::SpotBaseRESTAPI; pair::Union{Vector{String},String,Nothing}=nothing)

Kraken Docs: https://docs.kraken.com/rest/#operation/getTickerInformation

Returns the actual ticker of a given pair (asset pair/symbol).

Authenticated client not required

Examples

julia> client = SpotBaseRESTAPI()
julia> println(get_ticker(client, pair="DOTUSD"))
Dict{String, Any}(
    "DOTUSD" => Dict{String, Any}(
        "v" => Any["73347.09060256", "831710.16212826"], 
        "c" => Any["4.91490", "145.01800000"], 
        "o" => "4.88630", "t" => Any[635, 3486], 
        "b" => Any["4.91220", "265", "265.000"], 
        "l" => Any["4.82900", "4.82900"], 
        "a" => Any["4.91290", "82", "82.000"], 
        "p" => Any["4.89699", "4.97257"], 
        "h" => Any["4.93300", "5.07400"]
    )
)

KrakenEx.SpotMarketModule.get_tradable_asset_pair — Method

get_tradable_asset_pair(client::SpotBaseRESTAPI; pair::Vector{String}, info::String="info")

Kraken Docs: https://docs.kraken.com/rest/#operation/getTradableAssetPairs

Returns a list of available symbols (asset pairs) and additional informarion.

Authenticated client not required

Examples

julia> client = SpotBaseRESTAPI()
julia> println(get_tradable_asset_pair(client, pair=["XBTUSD"]))
Dict{String, Any}(
    "XXBTZUSD" => Dict{String, Any}(
        "margin_call" => 80, 
        "leverage_sell" => Any[2, 3, 4, 5], 
        "cost_decimals" => 5, 
        "lot_multiplier" => 1, 
        "long_position_limit" => 270, 
        "short_position_limit" => 180, 
        "fees_maker" => Any[
            Any[0, 0.16], 
            Any[50000, 0.14], 
            Any[100000, 0.12], 
            Any[250000, 0.1], 
            Any[500000, 0.08], 
            Any[1000000, 0.06], 
            Any[2500000, 0.04], 
            Any[5000000, 0.02], 
            Any[10000000, 0.0], 
            Any[100000000, 0.0]
        ], 
        "ordermin" => "0.0001", 
        "aclass_base" => "currency", 
        "fee_volume_currency" => "ZUSD", 
        "status" => "online", 
        "altname" => "XBTUSD", 
        "aclass_quote" => "currency", 
        "margin_stop" => 40, 
        "fees" => Any[
            Any[0, 0.26], 
            Any[50000, 0.24],
            Any[100000, 0.22], 
            Any[250000, 0.2],
            Any[500000, 0.18],
            Any[1000000, 0.16],
            Any[2500000, 0.14],
            Any[5000000, 0.12],
            Any[10000000, 0.1], 
            Any[100000000, 0.08]
        ], 
        "quote" => "ZUSD", 
        "base" => "XXBT",
        "lot" => "unit", 
        "pair_decimals" => 1, 
        "wsname" => "XBT/USD", 
        "costmin" => "0.5", 
        "tick_size" => "0.1", 
        "lot_decimals" => 8, 
        "leverage_buy" => Any[2, 3, 4, 5]
    )
)

User

KrakenEx.SpotUserModule.delete_export_report — Method

delete_export_report(client::SpotBaseRESTAPI; id::String, type::String)

Kraken Docs: https://docs.kraken.com/rest/#operation/removeExport

Removes an export by id from the system.

Authenticated client required

Attributes

type – must be one of ["delete", "cancel"]

...

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(delete_export_report(client, type="delete", id="PFEZ))
Dict{String, Any}("delete" => true)

KrakenEx.SpotUserModule.get_account_balance — Method

get_account_balance(client::SpotBaseRESTAPI)

Kraken Docs: https://docs.kraken.com/rest/#operation/getAccountBalance

Returns the actual balances of all currencies.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_account_balance(client))
Dict{String, Any}(
    "KFEE" => "9431.54", 
    "BCH" => "123.0000077100",
    "ZUSD" => "798.5491", 
    "XETH" => "14.0119368320", 
    "DOT" => "671.0153183000", 
    "ZEUR" => "101.12000", 
    "XXBT" => "1.0011888110", 
    "XXLM" => "2030121.00001212", 
    "EOS" => "24.0000065500", 
    "TRX" => "9121230.00000000"
)

KrakenEx.SpotUserModule.get_closed_orders — Method

get_closed_orders(client::SpotBaseRESTAPI;
    trades::Bool=false,
    userref::Union{Int64,Nothing}=nothing,
    start::Union{Int64,Nothing}=nothing,
    end_::Union{Int64,Nothing}=nothing,
    ofs::Union{Int64,String,Nothing}=nothing,
    closetime::String="both"
)

Kraken Docs: https://docs.kraken.com/rest/#operation/getClosedOrders

Returns information about the closed orders.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_closed_orders(client))
Dict{String, Any}(
    "closed" => Dict{String, Any}(
        "OLGRP5-M42MC-YOFF7T" => Dict{String, Any}(
            "price" => "0.00000", 
            "vol" => "40.36280000", 
            "status" => "canceled", 
            "vol_exec" => "0.00000000", 
            "oflags" => "fciq", 
            "reason" => "User requested", 
            "starttm" => 0, 
            "stopprice" => "0.00000", 
            "refid" => nothing, 
            "userref" => 0, 
            "expiretm" => 0,
            "cost" => "0.00000",
            "fee" => "0.00000", 
            "misc" => "", 
            "limitprice" => "0.00000", 
            "opentm" => 1.6730201392732968e9, 
            "descr" => Dict{String, Any}(
                "ordertype" => "limit",
                "price" => "40.5842", 
                "pair" => "DOTUSD", 
                "order" => "buy 40.36280000 DOTUSD @ limit 4.5842", 
                "leverage" => "none",
                "type" => "buy", 
                "close" => "", 
                "price2" => "0"
            ), 
            "closetm" => 1.673225209430317e9
        ), 
        "OUSQ7U-2TP3Q-PBKOIP" => Dict{String, Any}(
            ...
        )
    )
)

KrakenEx.SpotUserModule.get_export_report_status — Method

get_export_report_status(client::SpotBaseRESTAPI; report::String)

Kraken Docs: https://docs.kraken.com/rest/#operation/exportStatus

Returns information about the current export status.

Authenticated client required

Attributes

report – must be one of ["trades", "ledgers"]

...

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_export_report_status(client, report="ledgers"))
Any[Dict{String, Any}(
    "report" => "ledgers", 
    "fields" => "all", 
    "aclass" => "currency", 
    "endtm" => "1673368394", 
    "createdtm" => "1673368394", 
    "format" => "CSV", 
    "status" => "Queued", 
    "id" => "OFEZ", 
    "subtype" => "all", 
    "starttm" => "1672531200",
    "dataendtm" => "1673368394", 
    "completedtm" => "0", 
    "datastarttm" => "1672531200", 
    "expiretm" => "1674577994", 
    "flags" => "0", 
    "asset" => "all", 
    "descr" => "myLedgersExport"
)]

KrakenEx.SpotUserModule.get_ledgers — Method

get_ledgers(client::SpotBaseRESTAPI; id::Union{String,Vector{String}}, trades::Bool=false)

Kraken Docs: https://docs.kraken.com/rest/#operation/getLedgersInfo

Returns information about specific ledergs by id.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_ledgers(client, id="LNYQGU-SUR5U-UXTOWM"))
Dict{String, Any}(
    "LNYQGU-SUR5U-UXTOWM" => Dict{String, Any}(
        "amount" => "2001.0000",
        "balance" => "2001.0082", 
        "fee" => "0.0000", 
        "time" => 1.668681189483613e9, 
        "aclass" => "currency", 
        "asset" => "EUR", 
        "subtype" => "", 
        "refid" => "QYSCWYA-S54CSU-RDHRV3", 
        "type" => "deposit"
    )
)

KrakenEx.SpotUserModule.get_ledgers_info — Method

get_ledgers_info(client::SpotBaseRESTAPI;
    asset::Union{String,Vector{String}}="all",
    aclass::String="currency",
    type::String="all",
    start::Union{Int64,Nothing}=nothing,
    end_::Union{Int64,Nothing}=nothing,
    ofs::Union{Int64,String,Nothing}=nothing
)

Kraken Docs: https://docs.kraken.com/rest/#operation/getLedgers

Returns information about the current ledergs.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_ledgers_info(client))
Dict{String, Any}(
    "count" => 1068, 
    "ledger" => Dict{String, Any}(
        "LWSAYQ-GD2N2-IM3I6O" => Dict{String, Any}(
            "amount" => "0.00", 
            "balance" => "9437.94", 
            "fee" => "6.43", 
            "time" => 1.6732795570751212e9, 
            "aclass" => "currency", 
            "asset" => "KFEE", 
            "subtype" => "", 
            "refid" => "TYQEFQ-NYADC-R4LII6", 
            "type" => "trade"
        ), 
        "LP6VNL-HDXPP-JLVG73" => Dict{String, Any}(
            "amount" => "-0.0300000000", 
            "balance" => "0.0119368320", 
            "fee" => "0.0000000000", 
            "time" => 1.6732795570750895e9, 
            "aclass" => "currency", 
            "asset" => "XETH", 
            "subtype" => "", 
            "refid" => "TYQEFQ-NYADC-R4LII6", 
            "type" => "trade"
        ), ....
    )
)

KrakenEx.SpotUserModule.get_open_orders — Method

get_open_orders(client::SpotBaseRESTAPI; trades::Bool=false, userref::Union{Int64,Nothing}=nothing)

Kraken Docs: https://docs.kraken.com/rest/#operation/getOpenOrders

Retrurns the actual open orders.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_open_orders(client))
Dict{String, Any}(
    "open" => Dict{String, Any}(
        "OAQZLS-7SFKK-PAA6BW" => Dict{String, Any}(
            "price" => "0.00000", 
            "vol" => "20.11513967", 
            "status" => "open", 
            "vol_exec" => "0.00000000", 
            "oflags" => "fciq", 
            "starttm" => 0, 
            "stopprice" => "0.00000", 
            "refid" => nothing, 
            "userref" => 0, 
            "expiretm" => 0, 
            "cost" => "0.00000", 
            "fee" => "0.00000",
            "misc" => "", 
            "limitprice" => "0.00000", 
            "opentm" => 1.6678760796400564e9, 
            "descr" => Dict{String, Any}(
                "ordertype" => "limit", 
                "price" => "7.1288", 
                "pair" => "DOTUSD", 
                "order" => "sell 20.11513967 DOTUSD @ limit 7.1288", 
                "leverage" => "none", 
                "type" => "sell", 
                "close" => "", 
                "price2" => "0"
            )
        ), 
        "OPNPFB-NRZY3-O6T46Y" => Dict{String, Any}(
            "price" => "0.00000", 
            ...
        )
    )
)

KrakenEx.SpotUserModule.get_open_positions — Method

get_open_positions(client::SpotBaseRESTAPI;
    txid::Union{String,Vector{String},Nothing}=nothing,
    docalcs::Bool=false,
    consolidation::String="market"
)

Kraken Docs: https://docs.kraken.com/rest/#operation/getOpenPositions

Returns a list of open Margin positions.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_open_positions(client))
Dict{String,Any}(
    "ZBR4YTU-ND7TDJ-OZ8LR" => Dict{String,Any}(
        "ordertxid" => "03YHGT-P82PNP-D9DY85",
        "posstatus" => "open",
        "pair" => "XXBTZUSD",
        "time" => 1605280097.8294,
        "type" => "buy",
        "ordertype" => "limit",
        "cost" => "104610.52842",
        "fee" => "289.06565",
        "vol" => "8.82412861",
        "vol_closed" => "0.20200000",
        "margin" => "20922.10568",
        "value" => "258797.5",
        "net" => "+154186.9728",
        "terms" => "0.0100% per 4 hours",
        "rollovertm" => "1616672637",
        "misc" => "",
        "oflags" => ""
    )
)

KrakenEx.SpotUserModule.get_orders_info — Method

get_orders_info(client::SpotBaseRESTAPI;
    txid::Union{String,Vector{String}},
    trades::Bool=false,
    userref::Union{Int64,Nothing}=nothing
)

Kraken Docs: https://docs.kraken.com/rest/#tag/User-Data/operation/getOrdersInfo

Returns information about a specific order by txid.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_orders_info(client, txid="OLGRP5-M42MC-YOFF7T"))
Dict{String, Any}
    "OLGRP5-M42MC-YOFF7T" => Dict{String, Any}(
        "price" => "0.00000", 
        "vol" => "40.36280000", 
        "status" => "canceled", 
        "vol_exec" => "0.00000000", 
        "oflags" => "fciq", 
        "reason" => "User requested", 
        "starttm" => 0, 
        "stopprice" => "0.00000", 
        "refid" => nothing, 
        "userref" => 0, 
        "expiretm" => 0,
        "cost" => "0.00000",
        "fee" => "0.00000", 
        "misc" => "", 
        "limitprice" => "0.00000", 
        "opentm" => 1.6730201392732968e9, 
        "descr" => Dict{String, Any}(
            "ordertype" => "limit",
            "price" => "40.5842", 
            "pair" => "DOTUSD", 
            "order" => "buy 40.36280000 DOTUSD @ limit 4.5842", 
            "leverage" => "none",
            "type" => "buy", 
            "close" => "", 
            "price2" => "0"
        ), 
        "closetm" => 1.673225209430317e9
    )
)

KrakenEx.SpotUserModule.get_trade_balance — Method

get_trade_balance(client::SpotBaseRESTAPI; asset::Union{String,Nothing}=nothing)

Kraken Docs: https://docs.kraken.com/rest/#operation/getTradeBalance

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_trade_balance(client))
Dict{String, Any}(
    "v" => "0.0000", 
    "tb" => "113145.4781", 
    "mf" => "113145.4781", 
    "uv" => "0.0000", 
    "m" => "0.0000", 
    "c" => "0.0000", 
    "e" => "113112.4781",
    "eb" => "116412.4143", 
    "n" => "0.0000"
)
julia> println(get_trade_balance(client, asset="DOT"))
Dict{String, Any}(
    "v" => "0.0000000000", 
    "tb" => "2301.0859297648", 
    "mf" => "2301.0859297648", 
    "uv" => "0.0000000000", 
    "m" => "0.0000000000", 
    "c" => "0.0000000000", 
    "e" => "2301.0859297648", 
    "eb" => "2367.7876343538", 
    "n" => "0.0000000000"
)

KrakenEx.SpotUserModule.get_trade_volume — Method

get_trade_volume(client::SpotBaseRESTAPI; pair::Union{String,Vector{String},Nothing}=nothing, fee_info::Bool=true)

Kraken Docs: https://docs.kraken.com/rest/#operation/getTradeVolume

Returns the actual 30-day trading volume. Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_trade_volume(client))
Dict{String, Any}(
    "currency" => "ZUSD", 
    "volume" => "204212.9762",
    "fees_maker" => nothing, 
    "fees" => nothing
)

KrakenEx.SpotUserModule.get_trades_history — Method

get_trades_history(client::SpotBaseRESTAPI;
    type::String="all",
    trades::Bool=false,
    start::Union{Int64,Nothing}=nothing,
    end_::Union{Int64,Nothing}=nothing,
    ofs::Union{String,Int64,Nothing}=nothing
)

Kraken Docs: https://docs.kraken.com/rest/#operation/getTradeHistory

Returns historical trades and information.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_trades_history(client, type="all", start=1668431675, trades=true))
Dict{String, Any}(
    "count" => 182, 
    "trades" => Dict{String, Any}(
        "TRZUTT-EZZ7C-NYDTL3" => Dict{String, Any}(
            "price" => "4.29780", 
            "time" => 1.6722627506350572e9, 
            "ordertxid" => "OGFLW2-CZKU3-XKCBMB",
            "trade_id" => 8737419, 
            "vol" => "46.5350000", 
            "pair" => "DOTUSD", 
            "postxid" => "TKH2SE-M7IF5-CFI7LT", 
            "ordertype" => "limit", 
            "cost" => "199.9981", 
            "fee" => "0.3200", 
            "misc" => "", 
            "leverage" => "0", 
            "margin" => "0.00000",
            "type" => "buy"
        ), "TNJFKJ-YLYX7-ETYS66" => Dict{String, Any}(
            ...
        )
    )
)

KrakenEx.SpotUserModule.get_trades_info — Method

get_trades_info(client::SpotBaseRESTAPI; txid::Union{String,Vector{String}}, trades::Bool=false)

Kraken Docs: https://docs.kraken.com/rest/#operation/getTradesInfo

Returns information about specific trades by txid.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_trades_info(private_client, txid="2QMIXRF-I0EXWP-WQ6459"))
Dict{String,Any)(
    "2QMIXRF-I0EXWP-WQ6459" => Dict{String,Any}(
        "ordertxid" => "OQCLML-BW3P3-BUCMWZ",
        "postxid" => "0H90AJ-M4W3ZY-66GF8Y",
        "pair" => "XXBTZUSD",
        "time" => 1616667796.8802,
        "type" => "buy",
        "ordertype" => "limit",
        "price" => "30010.00000",
        "cost" => "600.20000",
        "fee" => "0.00000",
        "vol" => "0.02000000",
        "margin" => "0.00000",
        "misc" => ""
    )
)

KrakenEx.SpotUserModule.get_websockets_token — Method

get_websockets_token(client::SpotBaseRESTAPI)

Kraken Docs: https://docs.kraken.com/rest/#tag/Websockets-Authentication

Returns the websocket token. This is used by the when managing private websocket feeds and subscriptions.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_websockets_token(client))
Dict{String,Any}(
    "token": "1Agc3liasQPoAkAkadohsSNMFhs1ende06d1von3YFEMq",
    "expires": 900
)

KrakenEx.SpotUserModule.request_export_report — Method

request_export_report(client::SpotBaseRESTAPI;
    report::String,
    description::String,
    format::String="CSV",
    fields::String="all",
    starttm::Union{Int64,Nothing}=nothing,
    endtm::Union{Int64,Nothing}=nothing
)

Kraken Docs: https://docs.kraken.com/rest/#operation/addExport

Initiates the export of a report and returns the export-id.

Authenticated client required

Attributes

report – must be one of ["trades", "ledgers"]

...

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(request_export_report(
...        client, 
...        report="ledgers", 
...        description="myLedgersExport", 
...        format="CSV"
...    ) 
Dict{String, Any}("id" => "OFEZ")

KrakenEx.SpotUserModule.retrieve_export — Method

retrieve_export(client::SpotBaseRESTAPI; id::String)

Kraken Docs: https://docs.kraken.com/rest/#operation/retrieveExport

Retrieves the initiated exprort. Can be used to save this report to csv.

Authenticated client required

The id is obtained when requesting the report using request_export_report.

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> report = retrieve_export(private_client, id="OFEZ") 
julia> open("myExport.zip", "w") do io
...        write(io, report)
...    end

Trade

KrakenEx.SpotTradeModule.cancel_all_orders — Method

cancel_all_orders(client::SpotBaseRESTAPI)

Kraken Docs: https://docs.kraken.com/rest/#operation/cancelAllOrders

Cancels all open orders.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(cancel_all_orders(client))

KrakenEx.SpotTradeModule.cancel_all_orders_after_x — Method

cancel_all_orders_after_x(client::SpotBaseRESTAPI; timeout::Int)

Kraken Docs: https://docs.kraken.com/rest/#operation/cancelAllOrdersAfter

Cancels all open orders after timeout seconds. This can be used to avoid unwanted behavior in case of system instabilities. Set timeout to 0 to reset the timout.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(cancel_all_orders_after_x(client, timeout=60))
Dict{String, Any}(
    "currentTime" => "2023-01-10T17:01:57Z", 
    "triggerTime" => "2023-01-10T17:02:57Z"
)
julia> println(cancel_all_orders_after_x(client, timeout=0))
Dict{String, Any}(
    "currentTime" => "2023-01-10T17:01:59Z", 
    "triggerTime" => "0"
)

KrakenEx.SpotTradeModule.cancel_order — Method

cancel_order(client::SpotBaseRESTAPI; txid::String)

Kraken Docs: https://docs.kraken.com/rest/#operation/cancelOrder

Cancels an order by txid.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(cancel_order(client, txid="O2JLFP-VYFIW-35ZAAE"))

KrakenEx.SpotTradeModule.cancel_order_batch — Method

cancel_order_batch(client::SpotBaseRESTAPI; orders::Vector{String})

Kraken Docs: https://docs.kraken.com/rest/#operation/cancelOrderBatch

Enables the cancellation of multiple orders by txid.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(cancel_order_batch(
...        client, 
...        orders=[
...            "O2JLFP-VYFIW-35ZAAE", "O523KJ-DO4M2-KAT243", 
...            "OCDIAL-YC66C-DOF7HS", "OVFPZ2-DA2GV-VBFVVI"
...        ]
...    ))

KrakenEx.SpotTradeModule.create_order — Method

create_order(client::SpotBaseRESTAPI;
    ordertype::String,
    side::String,
    pair::String,
    volume::Union{Float64,Int64,String},
    price::Union{Float64,Int64,String,Nothing}=nothing,
    price2::Union{Float64,Int64,String,Nothing}=nothing,
    trigger::Union{String,Nothing}=nothing,
    leverage::Union{Float64,Int64,String,Nothing}=nothing,
    stp_type::String="cancel-newest",
    oflags::Union{String,Vector{String},Nothing}=nothing,
    timeinforce::Union{String,Nothing}=nothing,
    starttm::String="0",
    expiretim::String="0",
    close_ordertype::Union{String,Nothing}=nothing,
    close_price::Union{String,Nothing}=nothing,
    close_price2::Union{String,Nothing}=nothing,
    deadline::Union{String,Nothing}=nothing,
    validate::Bool=false,
    userref::Union{Int32,Nothing}=nothing
)

Kraken Docs: https://docs.kraken.com/rest/#operation/addOrder

Enables the Spot trading mechanism. Various parameters can be specified for an individual trading preference.

Authenticated client required

Arguments

validate::Bool=false – Simulates order placement if set to true

...

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(
...        create_order(
...            private_client,
...            ordertype="limit",
...            side="buy",
...            volume=0.3,
...            pair="XBTUSD",
...            price="17000",
...            oflags="post"
...        )
...    )
Dict{String,Any}(
    "descr" => Dict{String,Any}(
        "order" => "buy 0.30000000 XBTUSD @ limit 17000.0",
    ),
    "txid" => ["0ZH90J-M4W3ZY-66GF8Y"]
)
julia> println(
...        create_order(
...            private_client,
...            ordertype="market",
...            side="buy",
...            volume=0.001,
...            pair="XBTUSD",
...            validate=true
...        )
...    )
Dict{String,Any}(
    "descr" => Dict{String,Any}(
        "order" => "buy 0.00100000 XBTUSD @ market",
    ),
    "txid" => ["T4HSG14-TYRCOC-7SWPQ0"]
)

KrakenEx.SpotTradeModule.create_order_batch — Method

create_order_batch(client::SpotBaseRESTAPI;
    orders::Vector{Dict{String,Any}},
    pair::String,
    deadline::Union{String,Nothing}=nothing,
    validate::Bool=false
)

Kraken Docs: https://docs.kraken.com/rest/#operation/addOrderBatch

Enables placing multiple orders and order-specific actions at once.

Authenticated client required

The validate key can be submitted to orders to simulate the result (see create_order).

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(
...        create_order_batch(
...            client,
...            orders=[
...                Dict{String,Any}(
...                    "close" => Dict{String,Any}(
...                        "ordertype" => "stop-loss-limit",
...                       "price" => 1000,
...                       "price2" => 900
...                   ),
...                   "ordertype" => "limit",
...                   "price" => 40000,
...                   "timeinforce" => "GTC",
...                   "type" => "buy",
...                   "userref" => Int32(123),
...                   "volume" => 2
...                ),
...                Dict{String,Any}(
...                    "ordertype" => "limit",
...                    "price" => 42000,
...                    "starttm" => "1668455555",
...                    "timeinforce" => "GTC",
...                    "type" => "sell",
...                    "userref" => Int32(999),
...                    "volume" => 5
...                ),
...                Dict{String,Any}(
...                    "ordertype" => "market",
...                    "volume" => 2,
...                    "type" => "buy"
...                ),
...                Dict{String,Any}(
...                    "ordertype" => "limit",
...                    "price" => 43000,
...                    "starttm" => "1668455555",
...                    "timeinforce" => "GTC",
...                    "type" => "sell",
...                    "userref" => Int32(999),
...                    "volume" => 5
...                )
...            ],
...            pair="XBTUSD",
...        )
...    )
...)
Dict{String, Any}(
    "orders" => Any[
        Dict{String, Any}(
            "descr" => Dict{String, Any}(
                "order" => "buy 2.00000000 XBTUSD @ limit 40000.0", 
                "close" => "close position @ stop loss 1000.0 -> limit 900.0",
                "txid" => "N5F2Y6-E898J5-8Y6QE0"
            )
        ), 
        Dict{String, Any}(
            "descr" => Dict{String, Any}(
                "order" => "sell 5.00000000 XBTUSD @ limit 42000.0",
                "txid" => "3YHGOT-P82PNP-D9DY85"
            )
        ), 
        Dict{String, Any}(
            "descr" => Dict{String, Any}(
                "order" => "buy 2.00000000 XBTUSD @ market",
                "txid" => "IZXJ1O-9HER70-XJ0PCD"
            )
        ), 
        Dict{String, Any}(
            "descr" => Dict{String, Any}(
                "order" => "sell 5.00000000 XBTUSD @ limit 43000.0",
                "txid" => "3YHGOT-P82PNP-D9DY85"
            )
        )
    ]
)

KrakenEx.SpotTradeModule.edit_order — Method

edit_order(client::SpotBaseRESTAPI;
    txid::String,
    pair::String,
    volume::Union{String,Int64,Float64,Nothing}=nothing,
    price::Union{String,Int64,Float64,Nothing}=nothing,
    price2::Union{String,Int64,Float64,Nothing}=nothing,
    oflags::Union{String,Vector{String},Nothing}=nothing,
    deadline::Union{String,Nothing}=nothing,
    cancel_response::Bool=false,
    validate::Bool=false,
    userref::Union{Int32,Nothing}=nothing
)

Kraken Docs: https://docs.kraken.com/rest/#operation/editOrder

Enables the editing of a placed order by txid.

Authenticated client required

Arguments

validate::Bool=false – Simulates order placement if set to true ...

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(
...        edit_order(
...            client,
...            txid="O2JLFP-VYFIW-35ZAAE",
...            pair="XBTUSD",
...            volume=4.2,
...            price=17000
...        )
...    )
Dict{String, Any}(
    "status" => "ok", 
    "originaltxid" => "O2JLFP-VYFIW-35ZAAE", 
    "orders_cancelled" => 0, 
    "descr" => Dict{String, Any}("order" => "Order edited")
)

Funding

KrakenEx.SpotFundingModule.cancel_withdraw — Method

cancel_withdraw(client::SpotBaseRESTAPI; asset::String, refid::String)

Kraken Docs: https://docs.kraken.com/rest/#operation/cancelWithdrawal

Cancels a withdrawal by refid. The refid returned by withdraw_funds when initiating a withdraw. Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> cancel_withdraw(client, asset="DOT", refid="VGBNO6T-KFPTAQ-LAKGA6")
true

KrakenEx.SpotFundingModule.get_deposit_address — Method

get_deposit_address(
    client::SpotBaseRESTAPI; 
    asset::String, 
    method::String, 
    new::Bool=false
)

Kraken Docs: https://docs.kraken.com/rest/#operation/getDepositAddresses

Returns the available deposit adresses of a given asset (symbol/currency pair).

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_deposit_address(client, asset="DOT", method="Polkadot"))
Any[
    Dict{String, Any}(
        "new" => true, 
        "expiretm" => "0", 
        "address" => "14MvdTudtJppS4WrUXg5mQiQiq6baUXt3KwhxxEDXnB7nMYv"
    ), Dict{String, Any}(
        "new" => true, 
        "expiretm" => "0", 
        "address" => "15mK6NDReFAxkP6menxPRZdGNzRJY31vDMe9uSUPWfZj64Vy"
    ), ...
]

KrakenEx.SpotFundingModule.get_deposit_methods — Method

get_deposit_methods(client::SpotBaseRESTAPI; asset::String)

Kraken Docs: https://docs.kraken.com/rest/#operation/getDepositMethods

Returns the available deposit methods for a given asset (symbol/currency pair).

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> get_deposit_methods(client, asset="DOT")
Any[Dict{String, Any}(
    "gen-address" => true, 
    "method" => 
    "Polkadot", 
    "limit" => false
)]

KrakenEx.SpotFundingModule.get_recend_deposits_status — Method

get_recend_deposits_status(
    client::SpotBaseRESTAPI;
    asset::String,
    method::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.kraken.com/rest/#operation/getStatusRecentDeposits

Returns the status of the recend deposit.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(
...        get_recend_deposits_status(
...             client, asset="DOT", method="Polkadot"
...        )
...    )
Any[]

KrakenEx.SpotFundingModule.get_recend_withdraw_status — Method

get_recend_withdraw_status(
    client::SpotBaseRESTAPI; 
    asset::String,
    method::Union{String,Nothing}=nothing
)

Kraken Docs: https://docs.kraken.com/rest/#operation/getStatusRecentWithdrawals

Returns information about the recend withdrawal.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_recend_withdraw_status(client, asset="DOT"))
Dict{String,Any}(
    "method": =>"Bitcoin",
    "aclass" => "currency",
    "asset" => "XXBT",
    "refid" => "AGBZNBO-5P2XSB-RFVF6J",
    "txid" => "THVRQM-33VKH-UCI7BS",
    "info" => "mzp6yUVMRxfasyfwzTZjjy38dHqMX7Z3GR",
    "amount" => "0.72485000",
    "fee" => "0.00015000",
    "time" => 1617014586,
    "status" => "Pending"
)

KrakenEx.SpotFundingModule.get_withdrawal_info — Method

get_withdrawal_info(
    client::SpotBaseRESTAPI;
    asset::String, 
    key::String, 
    amount::Union{Float64,Int64,String}
)

Kraken Docs: https://docs.kraken.com/rest/#operation/getWithdrawalInformation

Returns information about a specific withdrawal. The key can be the refid returned by withdraw_funds.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_withdrawal_info(
...        client, 
...        asset="DOT", 
...        key="pwallet1", 
...        amount="200"
...    ))
)
Dict{String,Any}(
    "method" => "Bitcoin",
    "limit" => "332.00956139",
    "amount" => "0.72485000",
    "fee" => "0.00015000"
)

KrakenEx.SpotFundingModule.wallet_transfer — Method

wallet_transfer(
    client::SpotBaseRESTAPI;
    asset::String,
    from::String,
    to::String,
    amount::Union{Float64,Int64,String}
)

Kraken Docs: https://docs.kraken.com/rest/#operation/walletTransfer

Enables the transfer of assets between different wallets e.g. from Spot to Futures wallet.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(wallet_transfer(
...        client, 
...        asset="XLM", 
...        from="Spot Wallet", 
...        to="Futures Wallet", 
...        amount=200
...    ))
Dict{String,Any}("refid" => "AJGLAE5-KSKMGR4-VPNPNV")

KrakenEx.SpotFundingModule.withdraw_funds — Method

withdraw_funds(
    client::SpotBaseRESTAPI;
    asset::String,
    key::String,
    amount::Union{Float64,Int64,String}
)

Kraken Docs: https://docs.kraken.com/rest/#operation/withdrawFunds

Initiates a withdraw and returns the refid to track the status.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(withdraw_funds(
...         client, 
...         asset="DOT", 
...         key="some-widthdrawal-id-from-gui", 
...         amount=200
...     ))
Dict{String,Any}("refid" => "AGBSO6T-UFMTTQ-I7KGS6")

Staking

KrakenEx.SpotStakingModule.get_pending_staking_transactions — Method

get_pending_staking_transactions(client::SpotBaseRESTAPI)

Kraken Docs: https://docs.kraken.com/rest/#operation/getStakingPendingDeposits

Returns a list of pending staking transactions of this user.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(get_pending_staking_transactions(client))
Any[Dict{String,Any}(
    "method" => "polkadot-staked",
    "aclass" => "currency",
    "asset" => "DOT.S",
    "refid" => "CN1EVQO-MOWU97-89F1D5",
    "amount" => "2010",
    "fee" => "0.01",
    "time" => 1622988877,
    "status" => "Initial",
    "type" => "bonding"
), ...]

KrakenEx.SpotStakingModule.list_stakeable_assets — Method

list_stakeable_assets(client::SpotBaseRESTAPI)

Kraken Docs: https://docs.kraken.com/rest/#operation/getStakingAssetInfo

Lists the available assets to stake.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(list_stakeable_assets(client))
Any[Dict{String,Any}(
    "method" => "polkadot-staked",
    "asset" => "DOT",
    "staking_asset": "DOT.S",
    "rewards" => Dict{String,ANy}()
        "reward" => "12.00",
        "type" => "percentage"
    ),
    "on_chain" => true,
    "can_stake" => true,
    "can_unstake" => true,
    "minimum_amount" => Dict{String,Any}(
        "staking" => "0.0000000000",
        "unstaking" => "0.0000000000"
    )
), ...]

KrakenEx.SpotStakingModule.list_staking_transactions — Method

list_staking_transactions(client::SpotBaseRESTAPI)

Kraken Docs: https://docs.kraken.com/rest/#operation/getStakingTransactions

Returns the list of staking transactions for this user.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(list_staking_transactions(client))
Any[Dict{String,Any}(
    "method" => "xbt-staked",
    "aclass" => "currency",
    "asset" => "XBT.M",
    "refid" => "2BPOYNJ-Q3P855-5B9BL4",
    "amount" => "0.612",
    "fee" => "0.0000000000",
    "time" => 1623132481,
    "status" => "Success",
    "type" => "bonding",
    "bond_start" => 1623132481,
    "bond_end" => 1623132481
), ...]

KrakenEx.SpotStakingModule.stake_asset — Method

stake_asset(client::SpotBaseRESTAPI; asset::String, amount::Union{Float64,Int64,String}, method::String)

Kraken Docs: https://docs.kraken.com/rest/#operation/stake

Initiates a staking action for a desired asset.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(stake_asset(client, asset="DOT", amount=2000, method="polkadot-staked"))
Dict{String,Any}("refid" => "ABFPN4-KSKVP4-VPNALT")

KrakenEx.SpotStakingModule.unstake_asset — Method

unstake_asset(client::SpotBaseRESTAPI; asset::String, amount::Union{Float64,Int64,String})

Kraken Docs: https://docs.kraken.com/rest/#operation/unstake

Enables the unstaking of a staked asset.

Authenticated client required

Examples

julia> client = SpotBaseRESTAPI(key="api-key", secret="secret-key")
julia> println(unstake_asset(client, asset="DOT", amount=2000))
Dict{String,Any}("refid" => "GT01XVK-1VCDWO-7HR18B")