Kraken Exceptions

This package uses custom exceptions which can be thrown or even catched. Exceptions are thrown when the responses include errors.

KrakenEx.ExceptionsModule.KrakenApiLimitExceededError — Type

KrakenApiLimitExceededError <: KrakenException{Exception}

API rate limit exceeded. Please check your rate limits.

KrakenEx.ExceptionsModule.KrakenAuthenticationError — Type

KrakenAuthenticationError <: KrakenException{Exception}

Credentials are invalid.

This can caused by invalid API keys or invalid request payload so that the encryption creates invalid payloads.

KrakenEx.ExceptionsModule.KrakenCannotOpenPositionError — Type

KrakenCannotOpenPositionError <: KrakenException{Exception}

User/tier is ineligible for margin trading.

KrakenEx.ExceptionsModule.KrakenCostMinimumNotMetError — Type

KrakenCostMinimumNotMetError <: KrakenException{Exception}

Cost (price * volume) does not meet costmin.

KrakenEx.ExceptionsModule.KrakenDeadlineElapsedError — Type

KrakenDeadlineElapsedError <: KrakenException{Exception}

The request timed out according to the default or specified deadline.

KrakenEx.ExceptionsModule.KrakenException — Type

KrakenException{T}

Base type for all custom exceptions of this package.

KrakenEx.ExceptionsModule.KrakenInsufficientAvailableFundsError — Type

KrakenInsufficientAvailableFundsError <: KrakenException{Exception}

Client does not have the necessary funds.

KrakenEx.ExceptionsModule.KrakenInsufficientFundsError — Type

KrakenInsufficientFundsError <: KrakenException{Exception}

Client does not have the necessary funds.

KrakenEx.ExceptionsModule.KrakenInsufficientMarginError — Type

KrakenInsufficientMarginError <: KrakenException{Exception}

Exchange does not have available funds for this margin trade.

KrakenEx.ExceptionsModule.KrakenInvalidAPIKeyError — Type

KrakenInvalidAPIKeyError <: KrakenException{Exception}

An invalid API-Key header was supplied.

KrakenEx.ExceptionsModule.KrakenInvalidAccountError — Type

KrakenInvalidAccountError <: KrakenException{Exception}

The account is invalid.

KrakenEx.ExceptionsModule.KrakenInvalidAmountError — Type

KrakenInvalidAmountError <: KrakenException{Exception}

The specified amount is invalid.

KrakenEx.ExceptionsModule.KrakenInvalidArgumentsError — Type

KrakenInvalidArgumentsError <: KrakenException{Exception}

The request payload is malformed, incorrect or ambiguous.

KrakenEx.ExceptionsModule.KrakenInvalidArgumentsIndexUnavailableError — Type

KrakenInvalidArgumentsIndexUnavailableError <: KrakenException{Exception}

Index pricing is unavailable for stop/profit orders on this pair.

KrakenEx.ExceptionsModule.KrakenInvalidNonceError — Type

KrakenInvalidNonceError <: KrakenException{Exception}

An invalid nonce was supplied.

KrakenEx.ExceptionsModule.KrakenInvalidOrderError — Type

KrakenInvalidOrderError <: KrakenException{Exception}

Order is invalid.

KrakenEx.ExceptionsModule.KrakenInvalidPriceError — Type

KrakenInvalidPriceError <: KrakenException{Exception}

Price is invalid.

KrakenEx.ExceptionsModule.KrakenInvalidReferenceIdError — Type

KrakenInvalidReferenceIdError <: KrakenException{Exception}

The requested referece id is invalid.

KrakenEx.ExceptionsModule.KrakenInvalidSignatureError — Type

KrakenInvalidSignatureError <: KrakenException{Exception}

An invalid API-Sign header was supplied.

KrakenEx.ExceptionsModule.KrakenInvalidStakingMethodError — Type

KrakenInvalidStakingMethodError <: KrakenException{Exception}

The staking method is invalid.

KrakenEx.ExceptionsModule.KrakenInvalidUnitError — Type

KrakenInvalidUnitError <: KrakenException{Exception}

The specified unit is invalid.

KrakenEx.ExceptionsModule.KrakenMarginAllowedExceededError — Type

KrakenMarginAllowedExceededError <: KrakenException{Exception}

User has exceeded their margin allowance.

KrakenEx.ExceptionsModule.KrakenMarginLevelToLowError — Type

KrakenMarginLevelToLowError <: KrakenException{Exception}

Client has insufficient equity or collateral.

KrakenEx.ExceptionsModule.KrakenMarginPositionSizeExceededError — Type

KrakenMarginPositionSizeExceededError <: KrakenException{Exception}

Client would exceed the maximum position size for this pair.

KrakenEx.ExceptionsModule.KrakenMarketInOnlyCancelModeError — Type

KrakenMarketInOnlyCancelModeError <: KrakenException{Exception}

Request can't be made at this time. Please check system status.

KrakenEx.ExceptionsModule.KrakenMarketInOnlyPostModeError — Type

KrakenMarketInOnlyPostModeError <: KrakenException{Exception}

Request can't be made at this time. Please check system status.

KrakenEx.ExceptionsModule.KrakenNotFoundError — Type

KrakenNotFoundError <: KrakenException{Exception}

The resource is not found.

KrakenEx.ExceptionsModule.KrakenOrderForEditNotFoundError — Type

KrakenOrderForEditNotFoundError <: KrakenException{Exception}

The order for edit could not be found.

KrakenEx.ExceptionsModule.KrakenOrderLimitsExceededError — Type

KrakenOrderLimitsExceededError <: KrakenException{Exception}

Order limits exceeded. Please check your open orders limit.

KrakenEx.ExceptionsModule.KrakenOrderMinimumNotMetError — Type

KrakenOrderMinimumNotMetError <: KrakenException{Exception}

Order size does not meet ordermin.

KrakenEx.ExceptionsModule.KrakenPermissionDeniedError — Type

KrakenPermissionDeniedError <: KrakenException{Exception}

API key doesn't have permission to make this request.

KrakenEx.ExceptionsModule.KrakenPositionLimitExceeded — Type

KrakenPositionLimitExceeded <: KrakenException{Exception}

Position limit exceeded. Please check your limits.

KrakenEx.ExceptionsModule.KrakenRateLimitExceededError — Type

KrakenRateLimitExceededError <: KrakenException{Exception}

API rate limit exceeded. Please check your rate limits.

KrakenEx.ExceptionsModule.KrakenServiceUnavailableError — Type

KrakenServiceUnavailableError <: KrakenException{Exception}

The matching engine or API is offline.

KrakenEx.ExceptionsModule.KrakenTickSizeInvalidCheckError — Type

KrakenTickSizeInvalidCheckError <: KrakenException{Exception}

Price submitted is not a valid multiple of the pair's tick_size.

KrakenEx.ExceptionsModule.KrakenToManyAdressesError — Type

KrakenToManyAdressesError <: KrakenException{Exception}

To many adresses specified.

KrakenEx.ExceptionsModule.KrakenUnavailableError — Type

KrakenUnavailableError <: KrakenException{Exception}

The requested resource is unavailable.

KrakenEx.ExceptionsModule.KrakenUnknownAssetError — Type

KrakenUnknownAssetError <: KrakenException{Exception}

The asset is unknown.

KrakenEx.ExceptionsModule.KrakenUnknownAssetPairError — Type

KrakenUnknownAssetPairError <: KrakenException{Exception}

The asset pair is unknown.

KrakenEx.ExceptionsModule.KrakenUnknownOrderError — Type

KrakenUnknownOrderError <: KrakenException{Exception}

Order is unknown.

KrakenEx.ExceptionsModule.KrakenUnknownPositionError — Type

KrakenUnknownPositionError <: KrakenException{Exception}

Position is unknown.

KrakenEx.ExceptionsModule.KrakenUnknownReferenceIdError — Type

KrakenUnknownReferenceIdError <: KrakenException{Exception}

The requested referece id is unknown.

KrakenEx.ExceptionsModule.KrakenUnknownWithdrawKeyError — Type

KrakenUnknownWithdrawKeyError <: KrakenException{Exception}

The requested withdrawal key is unknown.

KrakenEx.ExceptionsModule.MaxReconnectError — Type

MaxReconnectError <: KrakenException{Exception}

To many reconnect tries.

KrakenEx.ExceptionsModule.get_exception — Method

get_exception(name::String)

Returns the exception type by name.